Fix indeterminate behavior in ft_split, remov GNL
ft_split made use of this assignment:
array[i] = fill_array(start, s, &i);
The value of i was incremented inside of the function, which meant that
the value of i in array[i] was indeterminate. It could be either the
value of i before fill_array() modified it, or the value after
fill_array() modified it.
The macbooks in school consistently used the incremented value of i,
while my computer at home always used the pre-incremented value of i.

I now increment the value of i in a location that should always be
determinate.
array[i++] = fill_array(start, s);

Also removed GNL since I'm not happy with the version I have currently.
Will add it back after rewriting.